requirements.txt.txt
streamlit
opencv-python=4.10.0
numpy
mahotas
Pillow